Refactoring. 3. Тестовое задание. Блок "Базовый синтаксис" [#505049 #476149]
