+++ src/test/java/ru/job4j/array/SimpleStringEncoderTest.java
@@ -1,6 +1,7 @@
 package ru.job4j.array;
 
 import org.junit.jupiter.api.Test;
+import ru.job4j.SimpleStringEncoder;
 
 import static org.assertj.core.api.Assertions.assertThat;
